fix: resolve form validation and data mapping issues
