+++ frontend/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "set PORT=5000 && next dev",
-    "build": "next build",
+    "build": "set NODE_OPTIONS=--max_old_space_size=8192 && next build",
     "build:analyze": "set ANALYZE=true && next build",
     "start": "npx serve@latest out",
     "lint": "eslint . --ext .ts,.tsx,.js,.jsx",
